+++ changelog.md
@@ -18,3 +18,7 @@
 * Remove redundant native-or-bluebird dependency
 * Update bluebird dependency
 * Update dev dependencies
+
+## Next
+
+* Only use `co.wrap` on generators